fastapi_back/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlmodel>=0.0.22
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0
python-multipart>=0.0.12
pydantic>=2.9.0
pydantic-settings>=2.5.0
pyyaml>=6.0.1
requests>=2.31.0

